Tagging for release of Python 2.4.3

+++ Modules/cjkcodecs/_codecs_hk.c
@@ -2,6 +2,7 @@
  * _codecs_hk.c: Codecs collection for encodings from Hong Kong
  *
  * Written by Hye-Shik Chang <perky@FreeBSD.org>
+ * $CJKCodecs: _codecs_hk.c,v 1.4 2004/07/18 04:44:27 perky Exp $
  */
 
 #define USING_IMPORTED_MAPS
@@ -31,7 +32,7 @@ ENCODER(big5hkscs)
 	while (inleft > 0) {
 		ucs4_t c = **inbuf;
 		DBCHAR code;
-		Py_ssize_t insize;
+		int insize;
 
 		if (c < 0x80) {
 			REQUIRE_OUTBUF(1)
